refactor(corpus): move accuracy_rich onto the shared loader
Same file format, different row shape and a genuinely different scoring rule
(chunk-sequence equality, not per-field), so only the line handling is shared
via parse_as. That was the fifth copy of "skip blanks and // comments".

Rich-text report lines verified identical against the pre-change baseline.
